Front-end #4: Corrección en links de base.html, admin.html debido a la modularización del endpoint admin; corrección en app.py del Front-end, endpoint contacto repetido
