+++ Frontend/templates/admin.html
@@ -16,7 +16,7 @@
                         </div>
                     {% endif %}
                 {% endwith %}
-                <form method="post" action="{{ url_for('admin') }}">
+                <form method="post" action="{{ url_for('adminBp.admin') }}">
                     <div class="mb-3">
                         <input type="email" name="email" class="form-control form-control-lg custom-form-control" placeholder="Email" required>
                     </div>
