+++ Frontend/templates/base.html
@@ -55,7 +55,7 @@
                     <a class="nav-link" href="{{url_for('contactoBp.contacto')}}">Contactenos</a>
                 </li>
                 <li class="nav-item">
-                    <a href="{{url_for('admin')}}" class="btn btn-primary ml-xl-4">Administracion</a>
+                    <a href="{{url_for('adminBp.admin')}}" class="btn btn-primary ml-xl-4">Administracion</a>
                 </li>
             </ul>
         </div>
